requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
ItsDangerous==2.1.2
Click==8.1.7
Flask-Login==0.6.2
Flask-Bcrypt==1.0.1
Flask-WTF==1.2.1
mongoengine==0.27.0
flask-mongoengine==1.0.0
pymongo==4.6.0
dnspython==2.4.2
python-dotenv==1.0.0
google-generativeai
pip-system-certs
sentence-transformers
scikit-learn
pymupdf
python-docx
numpy
torch --index-url https://download.pytorch.org/whl/cpu
